tests/wpt/tests/css/css-align/parsing/justify-tracks-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Grid Level 3: getComputedStyle().justifyContent</title>
<link rel="help" href="https://drafts.csswg.org/css-grid-3/#tracks-alignment">
<meta name="assert" content="justify-tracks computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("justify-tracks", "normal");

test_computed_value("justify-tracks", "space-between");
test_computed_value("justify-tracks", "space-around");
test_computed_value("justify-tracks", "space-evenly");
test_computed_value("justify-tracks", "stretch");

test_computed_value("justify-tracks", "center");
test_computed_value("justify-tracks", "start");
test_computed_value("justify-tracks", "end");
test_computed_value("justify-tracks", "flex-start");
test_computed_value("justify-tracks", "flex-end");
test_computed_value("justify-tracks", "unsafe end");
test_computed_value("justify-tracks", "safe flex-start");
test_computed_value("justify-tracks", "left");
test_computed_value("justify-tracks", "unsafe right");

test_computed_value("justify-tracks", "normal, normal", "normal")
</script>
</body>
</html>
